fix(frontend): extract sync type constants for HMR compatibility
Move SYNC_TYPES and HISTORICAL_SYNC_TYPES to separate syncTypes.ts file.
Vite Fast Refresh requires component files to only export React components.
