+++ frontend/src/components/admin/SyncTab.tsx
@@ -3,15 +3,7 @@ import { format } from 'date-fns';
 import {
   Play,
   Loader2,
-  Calendar,
-  Users,
-  User,
-  Home,
-  Layout,
-  UserCheck,
-  FileText,
   Clock3,
-  Brain,
   Zap,
   RefreshCw,
   Settings2,
@@ -28,27 +20,7 @@ import { useProcessRequests } from '../../hooks/useProcessRequests';
 import { StatusIcon, formatDuration } from './ConfigInputs';
 import { clearCache } from '../../utils/queryClient';
 import ProcessRequestOptions, { type ProcessRequestOptionsState } from './ProcessRequestOptions';
-
-// Sync type configuration
-export const SYNC_TYPES = [
-  { id: 'sessions', name: 'Sessions', icon: Calendar, color: 'text-sky-600' },
-  { id: 'attendees', name: 'Attendees', icon: Users, color: 'text-emerald-600' },
-  { id: 'persons', name: 'Persons', icon: User, color: 'text-violet-600' },
-  { id: 'bunks', name: 'Bunks', icon: Home, color: 'text-amber-600' },
-  { id: 'bunk_plans', name: 'Bunk Plans', icon: Layout, color: 'text-rose-600' },
-  { id: 'bunk_assignments', name: 'Assignments', icon: UserCheck, color: 'text-indigo-600' },
-  { id: 'bunk_requests', name: 'Requests', icon: FileText, color: 'text-orange-600' },
-  { id: 'process_requests', name: 'Process', icon: Brain, color: 'text-teal-600' },
-] as const;
-
-export const HISTORICAL_SYNC_TYPES = [
-  { id: 'sessions', name: 'Sessions' },
-  { id: 'attendees', name: 'Attendees' },
-  { id: 'persons', name: 'Persons' },
-  { id: 'bunks', name: 'Bunks' },
-  { id: 'bunk_plans', name: 'Bunk Plans' },
-  { id: 'bunk_assignments', name: 'Assignments' },
-] as const;
+import { SYNC_TYPES, HISTORICAL_SYNC_TYPES } from './syncTypes';
 
 // Run all syncs mutation
 function useRunAllSyncs() {
